+++ doc/examples/fluidsimfoam-cbox/src/fluidsimfoam_cbox/output.py
@@ -3,7 +3,7 @@ from textwrap import dedent
 from inflection import underscore
 
 from fluidsimfoam.foam_input_files import DEFAULT_HEADER, FoamInputFile
-from fluidsimfoam.foam_input_files.blockmeshhelper import BlockMeshDict, Vertex
+from fluidsimfoam.foam_input_files.blockmesh import BlockMeshDict, Vertex
 from fluidsimfoam.output import Output
 
 code_control_dict_functions = dedent(