+++ doc/examples/fluidsimfoam-sed/src/fluidsimfoam_sed/output.py
@@ -4,7 +4,7 @@ import numpy as np
 from inflection import underscore
 
 from fluidsimfoam.foam_input_files import FoamInputFile
-from fluidsimfoam.foam_input_files.blockmeshhelper import BlockMeshDict
+from fluidsimfoam.foam_input_files.blockmesh import BlockMeshDict
 from fluidsimfoam.foam_input_files.fields import VolScalarField, VolVectorField
 from fluidsimfoam.output import Output