+++ src/fluidsimfoam/foam_input_files/__init__.py
@@ -14,8 +14,9 @@
 .. autosummary::
    :toctree:
 
-    blockmeshhelper
+    blockmesh
     fields
+    fv_schemes
 
 """