+++ doc/CHANGELOG.md
@@ -36,4 +36,4 @@ Security in case of vulnerabilities.
 - Can setup, run and reload simulations
 - Parser for OpenFOAM input files (:mod:`fluidsimfoam.foam_input_files`)
 - Utility to create blockMeshDict files
-  (:mod:`fluidsimfoam.foam_input_files.blockmeshhelper`)
+  (:mod:`fluidsimfoam.foam_input_files.blockmesh`)